requirements.txt
flask==2.0.1
opencv-python==4.11.0.86
mediapipe==0.10.21
numpy==1.26.4
python-dotenv==0.19.0
werkzeug==2.0.1
pyautogui==0.9.53
requests==2.31.0
pymongo==4.6.1
passlib==1.7.4
pyjwt==2.8.0
